feat: create new project detail page
